pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["silvanshine <theo.piacentini@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ipykernel = "^6.27.1"
flask = "^3.0.0"
scikit-learn = "^1.3.2"
ruff = "^0.1.8"
black = "^23.12.0"
nbqa = "^1.7.1"
mypy = "^1.7.1"
pymongo = {extras = ["srv"], version = "^4.6.1"}
python-dotenv = "^1.0.0"
icecream = "^2.1.3"
pandas = "^2.1.4"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"